pyproject.toml
[tool.poetry]
name = "contract-monitoring-bot"
version = "0.1.0"
description = ""
authors = ["ilapshin <igorlapshin@list.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
web3 = "^6.14.0"
pydantic = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
